fix: update loader version and improve mixin code structure
This commit updates the loader version in the gradle properties and enhances the structure of the RandomizableContainerBEMixin_LootRefiller class. The changes include refactoring the code for better readability and maintainability by removing unnecessary method definitions and optimizing imports. Additionally, the icon path in the fabric.mod.json file has fixed.

- Updated loader version from 0.15.11 to 0.16.0 for compatibility.
- Refactored mixin methods for clarity and removed unused code.
- fixed icon path in fabric.mod.json.